pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "feed"
version = "0.1.0"
description = "Personal newsletter intelligence agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=0.3.0",    # New Gemini SDK
    "feedparser>=6.0.0",      # RSS parsing
    "resend>=2.0.0",          # Email delivery
    "httpx>=0.27.0",          # HTTP client (used by anthropic)
    "beautifulsoup4>=4.12.0", # HTML parsing
    "lxml>=5.0.0",            # Fast HTML/XML parser
    "pyyaml>=6.0.0",          # Config files
    "pydantic>=2.0.0",        # Data validation
    "pydantic-settings>=2.0", # Environment config
    "typer>=0.12.0",          # CLI framework
    "rich>=13.0.0",           # Beautiful terminal output
    "python-dateutil>=2.9.0", # Date parsing
    "jinja2>=3.1.0",          # Email templates
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.40.0"]
gemini = ["google-genai>=0.3.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
]

[project.scripts]
feed = "src.cli:main"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = true
